Some minor utils refactor
