+++ muxtools/utils/__init__.py
@@ -1,9 +1,10 @@
-from . import convert, download, env, files, glob, log, parsing, types, subprogress, dataclass, ffprobe
+from . import convert, download, env, files, glob, log, parsing, probe, types, subprogress, dataclass, formats
 
 from .convert import *
 from .download import *
 from .files import *
-from .ffprobe import *
+from .probe import *
+from .formats import *
 from .env import *
 from .glob import *
 from .log import *
